move to a .templates.h
this avoid warnings with clang about undefined functions
